+++ src/tools/index.ts
@@ -35,3 +35,5 @@ export * from "./switchboard";
 export * from "./elfa_ai";
 export * from "./fluxbeam";
 export * from "./coingecko";
+export * from "./realm-governance";
+
